19 feature

+++ ark-net/Model/ArkTransaction.cs
@@ -7,8 +7,8 @@ namespace ArkNet.Model
         public string Id { get; set; }
         public int Type { get; set; }
         public int Timestamp { get; set; }
-        public BigInteger Amount { get; set; }
-        public BigInteger Fee { get; set; }
+        public long Amount { get; set; }
+        public long Fee { get; set; }
         public string SenderId { get; set; }
         public string TecipientId { get; set; }
         public string SenderPublicKey { get; set; }
